Rename sessionReplayUtils to sessionReplayConstants
