+++ packages/rum/src/types/sessionReplayConstants.ts
